vercel-app/api/requirements.txt
supabase==2.4.0
PyJWT==2.12.1
bcrypt==4.1.3
python-dotenv==1.2.2
